pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tiktok-research"
version = "0.1.0"
description = "Multi-run LLM annotation pipeline for health language analysis in social media"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "numpy>=1.26.2",
    "scipy>=1.11.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
